Update patch set 4
Patch Set 4:

Yea this stuff is getting abandoned,
i'll have to separate them and fix em up.
I need that longentropy typo to be fixed though. Is it possible to fix it on this qcom media? I read the media-caf was possible.
I tried it out with the display-caf and audio-caf from the new qcom4.4_2.7 branch. The media-caf breaks ADB and crashes the kernel at boot.
I'll have to update the media-caf. 
What are your recommendations?

Patch-set: 4
Label: Code-Review=0